Move CommandTrie to CommandTrie.cs
